Trivial refactorings
Use `StringBuilder` and `String.format` over concatenation for some performance
gain without much additional complexity.

Mark a lot of things final, but not things that we can expect to be
reassignable soon, even if they're not now.

Fixed the response-not-responce typos.

declare some configuration constants in _constants_. Namely the port number.