70-request-a-micro-loan
This commit includes tests and an update to the protocol doc, but for the
avoidance of doubt, these features are not implemented here:
- Social responsibility checks for loan-requests
- Any kind of story/justification to accompany the request
- Accepting requests
- Offering loans

I've used a slightly puzzling pattern in the LoanProposals implementation. I
wanted an approach that would be thread-safe for multiple users (relevant once
show-loans/requests and accept-request are implemented), and that would resemble
a DB query so that fewer changes would be needed once persistence is introduced.