Use 'chain-of-responsibility' pattern in Commands
This de-spaghettifies some commands, like RegisterCommand. And generally enables
code re-use and de-duplication.

I made every Command class have to implement `getSyntax` and added usage/help
messages for each command, implementing issue
`78-add-a-usage-help-for-each-command`.

I wasn't thrilled about the idea of using middleware that mutated the state of
the calling object, but seeing no better alternative, I used this approach to
set fields like `amount`, allowing middleware lower in the stack to rely on
them. Any suggestions as to how to avoid this are most welcome!