73-accept-a-microloan-request
When a request is granted, it is no longer current, and a new microloan is
created.

I've chosen `GRANTLOAN` as the command, over `ACCEPTLOANREQUEST`, which I think
is too verbose (and hard to read!). And to identify the request, each will have
an ID that is unique among active loan-requests, but which, in practice, will
just be the username of the requester (because at the moment users can only have
one active request at a time (and usernames are unique)).

For the avoidance of doubt, the following features have NOT been implemented:
- A `show-loan-requests` command (meaning that for a user to accept a request
they have to know about its existence through other means!)
- A `show-microloans` command (meaning that users aren't able to confirm that
the granted request has become a microloan!)
- Repayments of microloans. This will need some kind of integration with the
`pay-someone` functionality.
- Accruals of microloans. Obviously this would need to be a job that runs
periodically.